+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 ### Created
 
 - Add a SwiftUI Font provider in `DynamicFont`
+- Add scale parameter to `UIImage.ad_filled`.
 
 ## [11.3.0] - 2022-08-01Z